chore: move package-related functionality into an extension trait
